#10 GET /user/<username>を突っ込んでる
