+++ backend/src/routes/api/user/get_user.rs
@@ -1,0 +1,36 @@
+use serde::{Serialize, Deserialize};
+use rocket_contrib::json::Json;
+use chrono::NaiveDateTime;
+use crate::db::Connection;
+use crate::auth::Claims;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct UserResponse {
+    id: i32,
+    name: String,
+    avatar_url: Option<String>,
+    email: Option<String>,
+    created_at: NaiveDateTime
+}
+
+#[get("/user")]
+pub fn get(claims: Claims, conn: Connection) -> Json<Option<UserResponse>> {
+    use crate::models::{User, user::AsUser};
+    Json(claims.as_user(&conn)
+        .map(|User { id, name, avatar_url, email, created_at }: User| UserResponse {
+            id, name, avatar_url, email, created_at
+        })
+        .ok()
+    )
+}
+
+#[get("/user/<username>")]
+pub fn get_user_by_username(_claims: Claims, username: String, conn: Connection) -> Json<Option<UserResponse>> {
+    use crate::models::{User, traits::Find};
+    Json(User::find(username, &conn)
+        .map(|User { id, name, avatar_url, email, created_at }: User| UserResponse {
+            id, name, avatar_url, email, created_at
+        })
+        .ok()
+    )
+}
